requirements.txt
fastapi
uvicorn
sentence-transformers
pdfplumber
python-docx
scikit-learn